overlays: Fix icon pack target packages for AOSP
The new Pixel icon packs have their target launcher package set to Pixel
Launcher, which is not present in AOSP. This causes the icon packs to
not appear in ThemePicker.

The older Circular icon pack's ThemePicker overlay also has the wrong
target package, so fix that as well while we're at it.

Change-Id: I7699f07325b943938c30ee559f1477166698b5fa